Plot Eye Density

This function creates a plot of the eye density using ggplot2.

Usage

# S3 method for eye_density
plot(
  x,
  alpha = 0.8,
  bg_image = NULL,
  transform = c("identity", "sqroot", "curoot", "rank"),
  ...
)

Arguments

x

An "eye_density" object.

alpha

The opacity level for the raster layer (default: 0.8).

bg_image

An optional image file name to use as the background.

transform

The transformation to apply to the density values (default: c("identity", "sqroot", "curoot", "rank")).

...

Additional args

Value

A ggplot object representing the eye density plot.

See also

Examples

# Assume `ed` is an "eye_density" object
# Plot the eye density
plot_eye_density <- plot.eye_density(ed)
#> Error in plot.eye_density(ed): could not find function "plot.eye_density"